재고 임시 감소 로직 추가 (#33)
* chore: remove unnecessary

* chore: remove unnecessary

* feat: add redis repository

* feat: add deal stock reservation

* chore: add claude code

* chore: update claude.yml

* refactor: improve deal stock reservation with Lua script for atomicity


* chore: add claude.yml

* chore: remove unnecessary

* chore: remove unnecessary

* feat: add redis repository

* feat: add deal stock reservation

* chore: add claude code

* refactor: improve deal stock reservation with Lua script for atomicity


* refactor: implement atomic stock reservation with Lua script

- Add Lua script for atomic deal stock reservation
- Refactor saveAtomicInRedis to use Lua script execution
- Update confirmStock to remove only confirmed order reservation
- Add executeScript and decrement methods to RedisRepository
- Add deleteByOrderId to DealReservationRepository


* Add id-token permission to claude job

* feat: add redis repository

* feat: add deal stock reservation

* chore: add claude code

* refactor: improve deal stock reservation with Lua script for atomicity


* chore: add parameter

* feat: add decrease stock quantity version 2 (RDB)

---------

+++ common-module/src/main/java/dto/store/DealDetailResponse.java
@@ -14,7 +14,7 @@ public record DealDetailResponse(
         BigDecimal price,
         BigDecimal discountPrice,
         BigDecimal discountValue,
-        int stockQuantity,
+        long stockQuantity,
         DealStatus dealStatus,
         LocalDateTime pickupEndTime
 ) {
@@ -27,7 +27,7 @@ public record DealDetailResponse(
             BigDecimal price,
             BigDecimal discountPrice,
             BigDecimal discountValue,
-            int stockQuantity,
+            long stockQuantity,
             DealStatus dealStatus,
             LocalDateTime pickupEndTime
     ) {
