+++ store-service/src/main/java/com/leui/storeservice/domain/deal/entity/Deal.java
@@ -39,7 +39,7 @@ public class Deal extends BaseEntity {
     private BigDecimal price;
 
     @Column(nullable = false)
-    private int stockQuantity;
+    private long stockQuantity;
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
